refactor: 콘솔에서 move 입력시 입력값중 move 를 제거하도록 리펙토링
